refactor: Make Sidebar a shared component
