+++ src/app/chat/page.tsx
@@ -1,15 +1,19 @@
 "use client";
 
-import { auth } from "@/lib/firebase/initialize";
+import { auth, firestore } from "@/lib/firebase/initialize";
 import { FirebaseError } from "firebase/app";
 import { signOut } from "firebase/auth";
 import { useRouter } from "next/navigation";
-import Sidebar from "./SideBar";
+
 import { useUser } from "@/context/userContext";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+import { doc, getDoc } from "firebase/firestore";
+import fetchUserDoc from "@/lib/firebase/firestore/fetchUserDoc";
+import Sidebar from "../../components/SideBar/SideBar";
 
 export default function ChatPage() {
   const { user, isUserLoading } = useUser();
+  // const { userData, setUserData } = useState({});
   const route = useRouter();
 
   const router = useRouter();
@@ -19,23 +23,39 @@ export default function ChatPage() {
     }
   }, [isUserLoading, user, route]);
 
-  const handleSignOut = async () => {
-    console.log("sign out");
-    try {
-      await signOut(auth);
-      alert("Logout success");
-      router.push("/");
-    } catch (error) {
-      if (error instanceof FirebaseError) {
-        // const errorCode = error.code;
-        // const errorMessage = error.message;
-        // const errorMessage = getAuthErrorMsg(error);
-        alert(`Logout Error: ${error.message}`);
-      } else {
-        alert(`Logout Error: ${error}`);
-      }
+  // if (!user) {
+  //   alert("No user data provided");
+  //   return false;
+  // }
+  useEffect(() => {
+    if (user) {
+      fetchUserDoc(user);
     }
-  };
+  }, [user]);
+
+  // const fetchUserData = async () => {
+  //   if (!user) {
+  //     alert("No user data provided");
+  //     return false;
+  //   }
+  //   const userRef = doc(firestore, "users", user?.uid);
+  //   const docSnap = await getDoc(userRef);
+  //   try {
+  //     if (docSnap.exists()) {
+  //       console.log("Document data:", docSnap.data());
+  //       // setUserData({
+  //       //   ...userData,
+  //       //   nickname: userData.nickname,
+  //       //   imgUrl: userData["imageUrls"][0],
+  //       // });
+  //     } else {
+  //       // docSnap.data() will be undefined in this case
+  //       console.log("No such document!");
+  //     }
+  //   } catch (e) {
+  //     console.log("Error getting cached document:", e);
+  //   }
+  // };
 
   if (isUserLoading) {
     return <div>Loading...</div>;
@@ -44,14 +64,14 @@ export default function ChatPage() {
   return (
     <div>
       <Sidebar>
-        <button className="btn" onClick={handleSignOut}>
-          Log out
-        </button>
         {/* <button className="btn mt-2" onClick={() => router.push("/get-start")}>
           {"start (test)"}
         </button> */}
       </Sidebar>
-      <main className="ml-32 px-2">{/* <p>chat page</p> */}</main>
+      <main className="ml-32 px-2">
+        {/* <h3 className="text-lg">{userData.nickname}</h3> */}
+        {/* <p>chat page</p> */}
+      </main>
     </div>
   );
 }
